// MarkEd-front/src/api/models/NotificationSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type NotificationSchema = {
    id: number;
    subject: number;
    message: string;
    link: string;
    is_read: boolean;
    date: string;
    assignment_id: number;
    assignment_title: string;
};